requirements.txt
#Will also need to Download MEGA11 command Line version (Not working with the newer version MEGA12)
bcrypt==4.2.1
biopython==1.85
cffi==1.17.1
colorama==0.4.6
contourpy==1.3.1
cryptography==44.0.1
customtkinter==5.2.2
cycler==0.12.1
darkdetect==0.8.0
fonttools==4.56.0
geographiclib==2.0
geopy==2.4.1
kiwisolver==1.4.8
matplotlib==3.10.0
numpy==2.2.3
packaging==24.2
pandas==2.2.3
paramiko==3.5.1
pillow==11.1.0
pycparser==2.22
PyNaCl==1.5.0
pyparsing==3.2.1
python-dateutil==2.9.0.post0
pytz==2025.1
scp==0.15.0
six==1.17.0
tqdm==4.67.1
tzdata==2025.1
openpyxl==3.2.0b1
pip==25.0.1